Update mmu_erec_cutter.cfg
Updated variable_feed_length for EREC cutter to work with V2 out of the box, which is based on the gate_parking_distance from mmu_parameter.cfg and adjusted the comment to show the values for ERCF V1.1 and V2+ others. This is based on the install.sh gate_parking_distance of 13 (ERCF V2 + others) vs 23(ERCF V1.1).

If not adjusted properly the cut length is not correct.